// server/ThirdPartyServices/Podium/stateManager/model/StateProperties.ts
export enum StateProperties {
    accessToken = 'accessToken',
    state ='state',
    refreshToken ='refreshToken',
    authToken = 'authToken',
    lastRefresh = 'lastRefresh',
    webhooksRegistered = 'webhooksRegistered',
    initialContactSyncCompleted = 'initialContactSyncCompleted',
    organizationId = "organizationId",
    tagTracker = "tagTracker",
}